Move methods & setStates to App
